+++ core/src/Math.hpp
@@ -3,11 +3,11 @@
 #include <cmath>
 
 namespace models::math {
-  inline bool is_approx(long double a, long double b) {
+  inline bool is_approx(double a, double b) {
     return fabs(a - b) < 0.00001;
   }
 
-  inline bool is_module_approx(long double a, long double b) {
+  inline bool is_module_approx(double a, double b) {
     return is_approx(fabs(a), fabs(b));
   }
